fix: step2 - 도메인으로부터 뷰 관련 책임을 분리
